Close the prepared quartic retry loop and measure the remaining port gap
One native entry now derives all three collection and reconstruction passes
for x^4-20018*x-20034, accepting at 152 relations with exact source regulator.
CPython, JavaScript and GMP agree on complete relation, generator, logarithm
and HNF checkpoints. Capped prefixes remain unpublished and terminal retries
preserve every owner. Correct the composed log-count buffer ABI and rerun its
107-case helper suite. Retain explicit unsupported driver branches and the
prepared-input boundary; no fundamental-unit-map claim is made.

Freeze and record three alternating prepared cubic comparisons with matching
outer work. PARI takes about 3.08 ms; both native representations about 109 ms.
Owner reset costs another roughly 29 ms outside the kernel measurements.
The tagged graph has no whole-helper GMP bridges. This is a real remaining
gap requiring phase attribution, not competitive performance or full-panel
qualification. Preserve raw measurements, exact hashes, limits and boundaries.

+++ bench/pari-class-group-port/collector_next_pass.py
@@ -20,7 +20,7 @@ def pari_prepare_next_small_norm_pass(
     outer: Int64Buffer,
     cache: IntegerBuffer,
     schedule: Int64Buffer,
-    log_completed: Int64Buffer,
+    log_completed: IntegerBuffer,
 ) -> int:
     """Return 0 prepared, -1 requiring source cache reallocation.
